Add -r flag to set PR_SET_CHILD_SUBREAPER
This is mostly necessary for processes that do graceful reloads of
executables, such as caddy (and possibly nginx). Because caddy will
spawn a new executable and then the parent will die, this causes caddy
to be reparented to pid 1 (init). So, to keep the orphan process under
a process supervisor, it's necessary for either the supervisor to
implement subreaping (not always desirable, and creates a hard Linux
dependency) or to place a subreaper between the supervisor and the
child process.

Ordinarily I'd use izabera's pm-tools subreaper program for this, but
I can't have signals killing the subreaper itself. So, in this case,
it's necessary to trap all signals and foward them to the child process
instead of the subreaper handling them.

To support part of this, signals are now forwarded to the process group
instead of the direct child process, since it's necessary to send it to
reparented processes.

In addition, SIGCHLD is now always swallowed when it reaches retrap.

This is something that can be supported on other operating systems
(FreeBSD and DragonFly both have procctl and PROC_REAP_ACQUIRE), but
requires support from golang.org/x/sys/* for the system calls each
platform requires.